requirements.txt
flask==3.0.0
flask-cors==4.0.0
supabase==2.3.0
python-dotenv==1.0.0
httpx==0.24.1
gotrue==2.4.2
